Remove value accessor animation code and optimize `ComponenType` repr (#14078)
Removes the old `ValueAccessorReadWriteAnimate` and associated
animation-driver code. We could/should try to reintroduce something like
it to sandbox side, especially for RPCs where we only want to send event
edges over the network.

I also changed the `ComponentType` enum to `u32` instead of `u64`, which
saves 32-64 bits in some cases on frequently used FFI structs. It's hard
to tell what's noise in benchmarks but it was consistently slightly
faster than main, between 1-5% depending on the run (with the simple
profiler comparison) 🤷‍♀️