Redo animation blend layer storage and networking representation (#13976)
The goal of this change is to optimize the network representation of animation blend layers; specifically, to make it so the per-frame delta is only 3 bytes in the form of a `UNorm16` for the time and a `UNorm8` for the weight.

Doing so became more complex than I originally anticipated, but it resulted in an overall improvement anyway, so I think was worthwhile.

There's quite a lot of meaningless tiny changes here because I removed some networking in-between structs, so reviewing might be a bit more difficult, sorry!

https://github.com/EmbarkStudios/wim-app/pull/13976